UML Window is now scrollable
Added the ability to increase or decrease the size of the window
through a menu item or a short cut.

The UML window will increase its size when class boxes are loaded from
a file to ensure that class boxes placed on a larger screen can be
scrolled to when using a smaller-screen machine.

The UML window will increase its size if you resize the window, so
class boxes placed after resizing the window can be scrolled to if you
shrink the size of the window.